dac_buffer: De-duplicate sine LUT
sine_lut_2 is a 90 degree shifted version of sine_lut_1, which means it
contains the same data, but by an offset of 256 samples. This is used to
generate I and Q pairs of samples.

The same can be accomplished by using one table and adjusting the index
which is used to access the table.

Having only one table reduces the code size, makes it easier to update the
table and also avoids accidental mistakes where both tables could have
slightly different content.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>