ad9361: Make sure to go to ALERT state when feedback clock changes
When the feedback clock changes, either in phase or frequency, or any other
discontinuities appear in the feedback clock, the AD9361 must go to the
ALERT state.

Otherwise an incorrect internal state might be entered and data on the
transmit path might get corrupted. This issue is visible during interface
tuning where the TX tuning can fail under certain circumstances.

To avoid this issue make sure that the AD9361 is always in the ALERT state
when changing the feedback clock during interface tuning. This makes sure
that the interface tuning always succeeds if there is a valid delay setting
that results in a working interface.