adxl372: Make sure that FIFO data is not stored out of order
In order to ensure that data in the FIFO is not overwritten or stored out
of order, at least one sample set must be left in the FIFO after every read.

This is why before calling adxl372_get_fifo_xyz_data, the number of samples
to be read is decremented by 3, which represents an xyz sample set.

Also, when the FIFO is configured, it is not correct to decrement the
number of samples to be written in the FIFO_SAMPLES register. This caused
the z axis value of the last sample in the FIFO to be wrong.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>