spi_init(): Pass init structure by reference instead of by value
There is no need to make a copy of the init strcture when passing it to
spi_init(). Instead pass a pointer.

Results in smaller generated code size as well as runtime stack usage.

Size of the fmcdaq2 project

         text    data  bss    dec     hex
Before   107972  2660  23392  134024  20b88
After    107844  2660  23392  133896  20b08

Users updated automatically with the following coccinelle script
// <smpl>
@@
expression a, b;
@@
spi_init(a,
-b
+&b
 )
// </smpl>

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>