jesd_core: Pass state struct by reference instead of by value
There is no need to make a copy of the jesd_core state struct each time one
of the functions is called. Pass the state struct as a pointer.

This results in smaller generated code and less runtime stack usage. It
also allows functions to make changes to state struct.

Size of the fmcdaq2 project

         text    data  bss    dec     hex
Before   107644  2660  23392  133696  20a40
After    107412  2660  23392  133464  20958

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>