ad9144: Add DAC PLL support
The AD9144 features an internal PLL that can be used to generate the DAC
clock from a slower external reference.

Add support for configuring the PLL to the AD9144 driver.

Using the PLL is optional and it can be enabled by setting the 'pll_enable'
field in the ad9144_init_params struct to 1. When using the PLL the
pll_ref_frequency_khz must be set to the frequency of the external
reference clock frequency and the pll_dac_frequency_khz field must be set
to the target DAC output frequency of the PLL in kHz.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>