fmcdaq2: Add an example configuration for internal DAC PLL
Add an example configuration that shows how the internal PLL of the AD9144
can be used to generate a DAC clock that is faster that the maximum clock
output rate of the AD9523-1 (1 GHz).

In this example mode configure the DAC for a sampling rate of 2 GSPS. Since
the JESD204 link can only handle 1 GSPS of data also set the internal
interpolation factor to 2.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>