ad7124: update header to be self-sufficient
Driver headers should be built without errors regardless of the order in
which they are compiled. This is accomplished by using the minimum
neccesary set of '#include' directives in the header. In this case, stdint
and platform_drivers must be included in the header as the declaration of
several functions and the definintion of the structs use them.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>