Add helper macro for initializing SPI chip_select field
The chip_select field of the spi_init_param struct expects that all bits
are asserted except for the device that is being addressed. Doing this by
hand is quite tedious and also difficult to read.

Add a helper macro that creates a chip select mask based on the chip select
index that should be selected.

All users are updated.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>