ad9144: Fix PLL setup corner case
The bcount setting of the PLL needs to be 6 or larger for the PLL to lock.

The current logic might pick for certain PLL output frequency and reference
frequency combinations a value for bcount that is less than 6.

In this case double the bcount value as well as the reference divider
value to mitigate the issue.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>