scripts/xsdb: Stop and reset CPU during reprogramming
It seems there is a fair chance that when reprogramming the FPGA when the
CPU is running can result in lockup of the CPU that it can't recover from
other than a hard system reset.

This can for example happen if `make run` is executed multiple times
without resetting the board in between.

Make sure that the CPU is stopped and reset before the new bitstream is
loaded.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>