scripts/xsct.tcl: Return error when build was not successful
When building the sw.elf file the xsct.tcl script currently always returns
success. Regardless of whether it was able to build the project or not. The
`sdk projects -build` command unfortunately neither returns an error code
nor throws and exception.

Add a manually check to see if the sw.elf file was generated successfully
and if not return an error code.

This for example allows `make` to abort if something goes wrong and it will
not try to run targets that depend on the sw.elf having been build
successfully.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>