fmcdaq3: Use recommended DAC link startup sequence
The startup of the JESD204 link for the DAC on the FMCDAQ3 board
occasionally fails with the current software setup.

Change the initialization order to follow the recommended sequence for
DACs. First start the JESD204 link layer and the JESD204 PHY layer in the
FPGA and only after that enable the DAC itself.

No link startup failures have been observed anymore when using the
recommended link startup sequence.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>