spi_write_and_read(): XSpiPs: Mask invalid chip select bits
The Xilinx PS7 SPI controller only has 3 chip select bits and the register
that contains those chip select bits is shared with other control bits. The
mask that is passed to XSpiPs_SetSlaveSelect() is not validated and if it
contains set bits beyond the first 3 it can corrupt other bits in the
control register.

To avoid this mask the chip select mask explicitly before passing it to
XSpiPs_SetSlaveSelect().

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>