ad9144: Specify samplerate instead of lanerate
From an application point of view samplerate is the more interesting
parameter than lanerate. Especially since the lanerate not only depends on
the samplerate but also other link parameters.

Rework the driver so that it accepts the samplerate rather than the
lanerate and then computes the lanerate from the samplerate and the other
link parameters.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>