ad9144: short_pattern_test(): Shorten delay
ad9144_spi_check_status() will wait for up to 100ms to wait for the short
pattern to be successfully detected. This is much longer then required as
detecting whether the short pattern is being transmitted correctly or not.

Detection in hardware is almost instantly, so wait only for a minimal
amount of time.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>