Add ad9144-fmc-ebz project
Add a project that supports the:
	* AD9135-FMC-EBZ
	* AD9136-FMC-EBZ
	* AD9144-FMC-EBZ

In the default configuration the project expects a clock in the range from
70-250 MHz to be connected to the J1 connector on the board.

This clock signal will be forwarded to both the converter and the FPGA. The
converter will be setup in a way so that the lane rate is equal to 40x the
clock rate.

The design will measure the frequency of the external clock signal and
configure the JESD204 subsystem accordingly.

The bitstream should be build from the dac_fmc_ebz HDL project.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>