adxl372: Refactor the driver
This patch restructures the existing adxl372 driver by adding a module
for SPI, while the baseline module deals with the chip-logic.

This is a necessary step, as this driver should support in the future
a similar device which differs only in the type of interface used (I2C
instead of SPI).

Signed-off-by: Stefan Popa <stefan.popa@analog.com>