ad5755: fix negative extreme case rollover
The AD5755 coding is offset binary. If the range is one of the rangese
that allow negative voltage output and voltage value given is lower than
the negative full scale, the output will swing to positive full scale. A
safer behaviour is to clamp the output at the negative full scale
instead. This is done by changing the way code is interpretted in the
driver from unsigned to signed to take full advantage of the conditional
statements.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>