common_drivers:xcvr_core: update DRP functions
Update XCVR driver DRP functions according to the Linux driver:

https://github.com/analogdevicesinc/linux/blob/master/drivers/iio/jesd204/xilinx_transceiver.c#L70a

There are two reasons for this change:

Firstly, similar names and method in the drivers makes them easier to
support. Whenever an issue is found in one of them, the other can be
easily checked and updated if neccesary.

Secondly, this makes the functions compatible with the Ultrascale
support that future patches provide.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>