common_drivers:xcvr_core:xcvr_modules: add Ultrascale support for QPLL
The main difference between Ultrascale and the other 7 series drivers is
accounting for the greater speed and different division registers for
the clock. This means the driver has separate functions for Ultrascale
devices and main functions that differentiate between devices.

This change is also designed to work with the new DRP driver using the
implementation from the Linux drivers.

https://github.com/analogdevicesinc/linux/blob/master/drivers/iio/jesd204/xilinx_transceiver.c#L643

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>