common_drivers:xcvr_core: revert incorrect struct member name
The xcvr_core struc member lane_rate_kbps was changed to lane_rate_khz
to be identical with the one in the Linux driver, but this name is not
correct because lane rate is accurately measured in bits per second not
hertz. Change the name back to the original more correct version.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>