ad7606: Add basic support for AD7606 family of devices
The AD7606/AD7606-6/AD7606-4 is a 16-bit 8/6/4 channel Analog to
Digital Converter (ADC). It can accommodate ± 10V and ±5V true bipolar
inputs. The AD7605-4 is part of the same family, the main difference is
that it does not support oversampling.

This patch adds basic serial support for these devices.

Datasheet:
Link: https://www.analog.com/media/en/technical-documentation/data-sheets/ad7606_7606-6_7606-4.pdf
Link: https://www.analog.com/media/en/technical-documentation/data-sheets/ad7605-4.pdf

Signed-off-by: Stefan Popa <stefan.popa@analog.com>