drivers:ad717x: add support for AD4112
The AD4112 is almost identical to the AD4111, the only difference being
that the AD4112 does not have Open Wire Detection capabilities.

Adding support for this part amounts to mentioning it in the header
comments and separating the Open Wire Detection bit for the two devices.

More information about AD4112:
https://www.analog.com/media/en/technical-documentation/data-sheets/AD4112.pdf

For comparison purposes the datasheet for AD4111:
https://www.analog.com/media/en/technical-documentation/data-sheets/AD4111.pdf

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>