ad7768-1: Rework the driver in a generic way
The current driver implementation is dependent to the Xilinx platform as
it uses specific spi_engine functions. However, the driver should also
be able to interface with a microcontroller using a classic SPI.

Newer versions of spi_engine driver, implement a spi_write_and_read()
function which works like a classic SPI. This patch adds this function
to the spi_engine driver which is part of this project and calls it in
the ad7768-1 driver.

Also, the spi_eng_setup() function is renamed to the more generic
spi_init().

Signed-off-by: Stefan Popa <stefan.popa@analog.com>