drivers:ltc2312: Add support
The LTC2312-14 is a 14-bit or 12-bit, 500ksps, serial sampling A/D
converter that draws only 3.2mA from a single 3V or 5V supply.

This driver is a modified LTC legacy driver and adds support for both
the 14-bit and 12-bit versions. The change from the original is only the
addition of the no-OS style API (setup and remove functions).

Part datasheet:
https://www.analog.com/media/en/technical-documentation/data-sheets/231214fa.pdf

Original driver:
https://github.com/analogdevicesinc/Linduino/tree/master/LTSketchbook/libraries/LTC2315

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>