scripts: xstc.tcl: Set the heap and stack sizes for MICROBLAZE
For MICROBLAZE platforms most projects require an increase in HEAP size
for the application to function properly. Increase the size of the heap and
the stack on the MICROBLAZE platforms to match the default sizes of the
ZYNQ platforms.

Do this by passing the defined symbols _HEAP_SIZE and _STACK_SIZE to the
linker, defined at 0x2000 to match the ZYNQ default values.

Define the _HEAP_SIZE with 0x2000 and the _STACK_SIZE with 0x2000 to
match the ZYNQ default values.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>