projects/drivers/xilinx_platform: Wrong SlaveSel value in function call
When writing to SPI, some mask is applied to chip_select value. This seems
to be wrong.

Valid values for ARM platform: "slaves can be selected with values 0-2"
Valid values for Microblaze platform: SlaveMask is a 32-bit mask with a 1
in the bit position of the slave being selected

Change CS values accordingly.

chipSelectIndex valid values(1~8), combined with Valid values for
Microblaze platform (a 32-bit mask with a 1 in the bit position of the
slave being selected) results in another change:
spi_ad_desc->chip_select = spiSettings->chipSelectIndex - 1;

For example:
If spiSettings->chipSelectIndex = 1
then spi_ad_desc->chip_select = 0
XSpi_SetSlaveSelect will be called with 0x01 << 0, value corresponding to
SlaveMask parameter.

Signed-off-by: Cristian Pop <cristian.pop@analog.com>