drivers:ad5592/3r: initialize internal reference
The internal reference is a core feature for the AD5592R and AD5593R
devices. The driver must have the posibility to initialize the internal
reference so that applications can activate it or leaving it deactivated in
a transparent manner. This helps applications to deliver a more complete
out-of-the-box experience.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>