ad9172: Add new driver for AD9172
The AD9172 is a high performance, dual, 16-bit digital-to-analog converter
(DAC) that supports DAC sample rates to 12.6 GSPS
An API is already provided for this driver:
no-OS/drivers/ad917x/ad917x_api, derived from
https://github.com/analogdevicesinc/linux/tree/dev_ad917x/drivers/iio/frequ
ency/ad917x.
The only difference between them is the use of stdint types, in the derived
one. The API should be omitted from the review.
A link to the datasheet of the AD9172 can be found here:
www.analog.com/media/en/technical-documentation/data-sheets/AD9172.pdf

Signed-off-by: Cristian Pop <cristian.pop@analog.com>