ad713x_fmc: update initialization function
The initialization routine has changed extensively with the changes in
the silicon. The DEVICE_CONFIG, MPC_CONFIG and the ODR_VAL registers do
not need to be initialized any more, but the DEVICE_CONFIG_1 and FIR_BW_SEL
do. The CLKOUT output and reference gain correction now need to be
enabled and the wideband filter's bandwidth for all channels needs to be
initialized.

Also the initialization in the driver was initializing the
GPIOs for two devices instead of one. This is changed to make the driver
more scalable.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>