ad713x_fmc: update magnitude/phase clock delay
The magnitude/phase clock delay function in the silicon has changed from a
global setting for all the channels to a setting for each channel
individually. Update the function to enable the delay globally, to keep
backwards compatibility, and mark it as deprecated. Add a function to
change the delay on a per channel basis.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>