adv7511: move sources in the main folder
The rework includes getting rid of slightly different sources for every
platform and instead add support for all the platforms in the same sources
by using a "config.h" file and compiler switches. Move sources from random
platform folder to the main folder to better highlight the changes made.
These will be the sources updated to include support for the rest of the
plaforms.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>