projects:drivers:axi_dmac: make write/read available
Make the functions axi_dmac_read and axi_dmac_write available to users of
the driver to allow more customization than the transfer function provides.
To do this the funtion declarations are added to the header and the
register defines are moved to the header too to be used with the functions.
The util.h include is moved in the header as well because the register
defines use it.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>