projects:drivers:axi_dmac: change the flag member to register HAL
The different bits in the FLAGS register of the AXI DMAC are defined in
software as an enum, but the way the flags member of the descriptor is used
is OR-ing the different bits in it and writing it dirrectly to the
register. This implementation is closer to a Hardware Abstraction and would
make more sense to be explicitly a 32 bit value (because the register is 32
bits long) than an object of the enum type.

This change has ideally no actual functionality change, but it better fits
in the code and may avoid future errors with different platforms and
compilers.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>