drivers:platform:xilinx: fix options for AXI IIC
The AXI IIC driver functions XIic_Send and XIic_Recv use the options
argument to control if the transmission ends with a STOP bit or a REPEATED
START. To use the repeated start enum the platform driver must send a 1
value for repeated start and 0 for a stop bit. Fix the implementation to do
that.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>