Drivers: platform: modify parameter type
Modify parameter type of bytes_number from uint8_t to uint16_t in
function spi_write_and_read. Modify file in includes and in all platform
files.
Uint8_t data type is not enough to represent the size of a data block
(512 bytes) for reading and writing data in a SD card over SPI.
This modification doesn't affect the existing spi drivers because in the
implementation there are no constrains for this parameter to be uint8_t

Signed-off-by: Chindris <Mihail.Chindris@analog.com>