drivers:platform:xilinx: SPI transmit fix
When the option XSPIPS_FORCE_SSELECT_OPTION is used for the transfer, as is
the case here the CS line will be asserted as soon as it is picked via the
XSpiPs_SetSlaveSelect method. If the slave is not deselected after the
transfer is complete the CS line will remain asserted potentially creating
problems, for example with new transfers, or with devices that only
update registers on CS deassert.

Add this patch to deselect the CS and deassert the CS line after every
transfer.

Signed-off-by: Drimbarean <Andrei.Drimbarean@analog.com>