include:timer: add load value for timer
The load value usually represents the value at which the timer starts
counting. It is the most used value when using the timer to time arbitrary
time intervals. Usually the timer has limited optons of clock source and
slightly more, but still limited options of prescaling that clock to
aobtain the desired counting frequency. For that reason, after configuring
counting frequency, the user would choose a starting value to fine tune the
frequency at which the timer will trigger an interrupt. This way
arbitrary periods of time can be measured with very high accuracy.

Add value to the initialization routine and the handler
structure for easy tracking.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>