travis: generate .dox files
Currently, the drivers_page.dox and projects_page.dox are populated
manually with sections and links to different types of drivers/projects.

This patch replaces the .dox files corresponding to drivers/projects and
with gen_dox.sh script that generates these files on doxygen travis
build.

For example, when a new driver implementation is merged on master branch,
this approach takes care of updating the doxygen files accordingly.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>