lookup table based crc-8 implementation
This is code compatible with linux crc8.h although the implementation is
somewhat different.

This implementation is supposed to replace all crc-8 implementations in no-OS
drivers which are all duplicates of each other.

By comparing the performance of this new implementation with the old one on
an x86 machine and on large sample sizes, it seems the new implementation is
faster by a factor of 13.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>