drivers: freq: ad9528: check if hw reset enabled
Not all projects that use ad9528 driver have a gpio connected to the
reset pin of the device.

This patch adds a `hw_reset_en` parameter to the device initialization
structure that is used to verify if hardware reset can be performed.

Update adrv9009 project to use the latest feature of the driver.

Fixes: a223f817f40aed7a33ade8b472855f0632171f73

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>