altera: gpio: update driver
Add implementation for `gpio_remove` and `gpio_get_value`.

Move implementation of `gpio_direction_output` to `gpio_set_value`. The
current implementation only sets the gpio value, while the information
regarding the direction of the gpio is contained in the `base_address`.

The `gpio_set_value` is called inside `gpio_direction_output` to avoid
errors within the projects that are using these functions.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>