drivers: adc: ad9250: update driver
Add `register_data` pointer as parameter to the ad9250_read function.

This holds the information about the values read from the device.

Until now, the function was returning the value read, while the checks
through the entire driver were expecting a return code.
(SUCCESS/FAILURE).

This patch updates the read function and remove redundant checks based
on the previous implementation.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>