core clock calculation for adrv9009 som
Core clocks provided to the fpga are actually the div40 clocks
which need to be dynamically computed based on lanerate, which instead
comes from profile.

This change improves the behavior of adrv9009 som on no-OS in
the sense that now, 200 and 400 BW profiles work well. 100 BW
profile still has problems, but it's the same on linux, so it
will be addressed in a different commit.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>