tools:scripts:platform:aducm3029: Add HardFault_Handler_C
This change is done because when debuging and trying to reset the
target with the reset button, the default handler will be called and
the target will not reset.
With this change, when pressing the reset button while debuggin, the target
will reset and the program will start again and there is no need to upload
the program each time a reset is needed.
The print function is used to signal an actual HardFault is one occured will debugging.

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>