platform: xilinx: gpio: fix error handling
For the PL GPIO scenario, if an error occurs at initialization, the
memory is freed but the return code will be SUCCESS, since the switch
case statement will break.

This patch makes sure that if an error occurs, the code will jump to
the default case and return the FAILURE error code.

Fixes #659.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>