iio_axi_adc: custom get_sampling_frequency()
It is sometimes not possible to get the exact sampling frequency from
HDL registers. Application may know more about the number of lanes and
other parameters that affect the true sampling rate of an ADC.

By adding this function pointer to iio_axi_adc, the user may provide a
custom function that knows how to get the exact sampling rate if an axi_adc.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>