tools:scripts: Update makefile
In this commit the bigger change is to move from aducm.mk to
generic.mk all that can be shared with other makefiles. Mainly
the build rules and utilitary function for windows and linux
platforms.
Then xilinx.mk was created based on
https://github.com/analogdevicesinc/no-OS/blob/ed2871a4bcd38c3d53812d1ac9aee5c2d92317c1/tools/scripts/linux.mk
and
https://github.com/analogdevicesinc/no-OS/blob/ed2871a4bcd38c3d53812d1ac9aee5c2d92317c1/tools/scripts/windows.mk
with needed rules to compleat generic.mk with the xilinx specific
functionality.

Improvments added:
 - Same file will work for both linux and windows. New features
will be added in just one place
 - After creating a project make will build just updated files
 - Flag -j for paralel build is supported
 - Rule clean will delete only binary files
 - Rule clean_all will delete the build folder and libraries
 - Rule project_build will build the project the SDK way
 - Rule ra: clean_all then all

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>