libraries: iio: Add support for reg map access from iio_osciloscope
This moves the iiod protocol register read logic to iio.c instead of
implementing it on each iio device. These pointer can be set directly
with the function like spi_reg_read and spi_reg_write instead of adding
wrappers.
It is similar to:
https://github.com/analogdevicesinc/linux/blob/f081a37e7e8d29a43a4dbe2d49cc08d9f073c7c8/include/linux/iio/iio.h#L481
But instead of one function, two are used in order to not create
wrappers for each device.

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>