stm32 (stm32f4) support in makefiles
This adds the possibility to build a project for the STM32
microcontrollers using:

make PLATFORM=stm32

The family is detected automatically from file names.

The build/ directory is used only to store the .o, .elf
and other build artifacts and, unlike for xilinx and intel
projects, it is not used to store an SDK project. The reason
behind this is that STM32CubeIDE does not support project
creation from command line. Therefore the build is entirely
maintained by us using our build system.

A new rule was added specifically for this platform, which is:

make PLATORM=stm32 debug

This rule has as prerequisite that the shell PATH contains

OPENOCD_BIN
OPENOCD_SCRIPTS

environment variables. If they are properly configured, the
rule creates an .openocd and a .gdb config script, starts
an openocd server and starts a gdb debug session with the .elf.

Additionaly, like for other platforms, one can flash the chip
and run execution by

make PLATFORM=stm32 run

Signed-off-by: Darius Berghe <darius.berghe@analog.com>