tools:scripts: implement standalone rule
Running make standalone rule, an sdk project will be created and it
can be moved to anyplace without depending on noos.
The only exception for the moment are mqqt, mbedtls and fatfs
libraries which needs the noos path.
Modification to include directories where made to be relative to
the project sdk directory.

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>