make debug on aducm
CrossCore debugger keeps crashing on me so I'm adding this
alternative method for debugging on aducm platform.

This commit adds the possibility to use gdb in -tui mode by invoking:
	make debug

The toolchain does not currently have a gdb with -tui mode included.
Therefore, try to use the stock gdb first (fixed path, will only
work on linux) because it has a -tui mode.
If this fails, simply use the toolchain gdb.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>