iio: uniform string formatting/parsing
Previously, the user had to manually specify the formatting of
the input/output string which is bad. This can lead to differences
between linux/no-OS iio driver behaviour.

Not only that, but because the previous implementation abadoned the
val & val2 usage, some existing iio drivers in no-OS reimplement
computations which could've been easily and without error imported
from the linux implementation which uses val & val2.

Therefore, the introduced functions keep using the val & val2 logic
to have nearly identical computations on linux and no-OS and for
uniform string formatting/parsing.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>