drivers: ada4250: bandwidth reverse logic
Apparently the logic is reversed for this one. When in low BW mode, the BW pin should tied to AVDD and when in high BW mode, the BW pin should be tied to AVSS.
The low BW mode is actually
the standard response while the high BW mode is more of a BW boost feature.