projects:ad713x_fmcz: fix data read bug
The projct was returning always the same sample of data because the one
of the ADCs was used in a reserved mode of operation. The chip can only
function as a master on the data bus (generate SPI clock and ODR) or a
slave (accept SPI clock and ODR). The chip was used instead to generate
ODR, but accept the SPI clock which cannot work properly.

To accomodate this use a an AXI PWM generator to generate ODR at a
specified rate for both devices. Note: the ODR signal must be present
before chip initialization.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>