iio: Add trigger support
A trigger can be created similar to an iio_device.
Those triggers can be set in the init param of iio.
A trigger can provide enable or disable callbacks that
will be called when enabling or disabling a buffer.
When a trigger signal is received iio_trigger_notify
should be called with the trigger name.
Then, when calling iio_step devices for the triggered triggers
will be notified

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>