tools:scripts:*.mk sdkbuild rule
The addition of the sdkbuild rule that builds the
project with the corresponding SDK tools. Applicable
to the ADuCM3029, STM32, and Xilinx platforms.
The build_project.tcl was modified for using the
command recommended by Xilinx for dealing with
projects (app command). The addition of the
remove_dir_action Linux command in generic.mk
for assuring consistency with Win commands.

Signed-off-by: George Mois <george.mois@analog.com>