iio: Introduce iio_snprintf to harden printing to buffers
Quote from 'man snprintf':

The functions snprintf() and vsnprintf() do not write more than size bytes
(including the terminating null byte ('\0')). If the output was truncated due
to this limit then the return value is the number of characters (excluding the
terminating null byte) which would have been written to the final string if
enough space had been available. Thus, a return value of size or more means
that the output was truncated.

Signed-off-by: Alexander Vickberg <wickbergster@gmail.com>