iio: Add trigger support
A trigger can be created similar to an iio_device.
Those triggers can be set in the init param of iio.
A trigger can provide enable or disable callbacks that
will be called when enabling or disabling a buffer.
A trigger can be handled synchronously with the interrupt (in the
same context) or asynchronously, in the iio_step context.
When a trigger signal is received iio_process_trigger_type
should be called with the trigger name.
Then, based on the configuration of the trigger
(is_synchronous = true or is_synchronou = false) the handler will be
called either directly (if true) either when iio_step is executing
(when false).

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>