stm32 azure builds
Add possibility to build stm32 projects on azure.
This improves the script by also making it load the environment
variables for a particular platform dynamically instead of
relying on .bashrc. A modification on the build server is needed:

~/.xilinx_environment.sh
~/.aducm3029_environment.sh
~/.stm32_environment.sh

The above scripts need to exist and prepare the environment for
the particular platform.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>