projects:spi_engine_test: final tests of DUAL SPI
This is the final test of the SPI with two lanes. Although the tests
show that the AD3552R responds, it comes at a cost of zero-padding the
sent data, in order to send the address on one lane. This would work
but it's a bad solution, thus, for now, DUAL SPI will not be supported.

Signed-off-by: chegbeli <ciprian.hegbeli@analog.com>