include:no_os_irq.h: Added handle fields
For no_os_callback_desc, the handle field would store
extra information needed for callback registration. This
replaces the need to pass this info in the irq_ctrl_desc
struct, as that could be used for more that one peripheral.
Changing the extra field on each register_callback call
would lead to overwrite issues.

By convention, in the handle fields, the user has to pass
HAL descriptors.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>