drivers:dac:dac_demo: Adapt driver for writedev usage
made changes such that writedev functionality is working as
expected with loopback enabled.

writedev was not working as expected because write_dev callback
was not initialized with dac_write_samples (instead read_dev
was initialized with dac_write_samples, which is actually used
for reading data from the IIO server).
there was also an issue in dac_demo_init, because
loopback_buffer_len was not initialized, which is later used
erroneously for writing the data in dac_write_samples.
there was also an issue in dac_write_samples, when writing to
desc->loopback_buffers: the data was not written in the buffer.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>