drivers:platform:maxim:maxim_uart.c: Fix reinit
This change fixes a bug that occured when the UART descriptor
is removed and initialized again. The problem was that the
callbacks were not unregistered automatically, and registering
a new one in the init function would result in an error (this
situation happens when iio_app's UART desc would use a baudrate
other than UART_DEFAULT_BAUDRATE).

This is an improvised solution for chips that use the UART
RevB driver, and will be changed once the SDK bug is fixed.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>