drivers:adc:adc_demo: Modify adc_demo from 12bit to 16bit data
This change is performed such that both adc and dac devices
are using 16-bit format. It is useful when using loop-back
feature (the data written to the dac is transferred to the
buffer used by the adc).
Also, in case iio_readdev is used without loop-back feature,
the data will be stored in 2's complement with sign bit
on bit16, instead of bit12, allowing gnuplot to render
the data correctly.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>