stm32 build hang workaround
Due to a yet elusive problem, stm32cubemx sometimes hangs while building
on azure.

This commit adds a workaround to not block the other builds by
time bounding the make all execution.

It seems reverting stm32cubemx to 6.3.0 would fix the problem, but 6.3.0
has another issue that was the reason we currently  use 6.5.0.

So this workaround should be removed once 6.6.0 arrives and fixes the
code generation hanging problem.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>