cn0565 project prototype (stm32 & aducm3029)
This is the initial version of the cn0565 project for Body
Impedance Analysis.

The projects currently builds for and runs on aducm3029 and SDP-K1.

`make` builds a custom application that communicates over UART to a
corresponding libiio app (sw_eit_ad5940).

`make TINYIIOD=y` builds an app that runs an iiod server. The same
libiio app (sw_eit_ad5940) may be used in iio mode to connect to
the iiod server, but iio-oscilloscope or other iio clients may be
used as well.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>