iio:iio.c: Add compiler switch for buffer overrun error
In case a buffer overrun takes place, the iio will treat this
behavior as an error and will return an error to the iio client.
In some cases, the user might want to ignore this error,
even if it means overwriting sampled data which was not sent
to the client.
By adding IIO_IGNORE_BUFF_OVERRUN_ERR flag in
CFLAGS the user can now enable this functionality.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>