include:no_os_timer: Rename load_value to ticks_count
The name load_value does not describe the parameter usage
correctly. It implies that it is the start value of the timer,
however we use it to describe the number of ticks the timer
counts until it resets. A rename is done to better describe
this parameter usage.

No functional changes.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>