adi_axcvr.* xilinx_transceiver.c PLL selection
The modification of the mechanism for selecting PLL used for
achieving desired rate.

Addition and use of defines for selecting the PLL reference clock source
to drive RXOUTCLK. CPLL or QPLL are enabled depending on the selected
reference clock.

The update of all sample projects to use the PLL reference
clock selection.

Signed-off-by: George Mois <george.mois@analog.com>

src:app:fmcdaq2.c Update of reference clock selection

The update of the fmcdaq2 sample project to use the PLL reference
clock selection to drive the RXOUTCLK and to enable\disable CPLL or
QPLL depending on this selection.

Signed-off-by: George Mois <george.mois@analog.com>