stm32 spi: configure peripheral before transfer
Previously, the stm32_spi_write_and_read function did not
configure the spi peripheral with the settings stored in the
spi device descriptor. This means that it was impossible to use
the driver to drive multiple slaves with different spi settings (
bus speed, spi mode, bit order) on the same spi bus.

This commit changes the behavior of spi_write_and_read on stm32
platform to reconfigure the spi peripheral before each transfer if
needed.

The condition to assess the need for reconfiguration is based on
 - spi instance
 - chip select

If any of the above changes between consecutive transfers, the
reconfiguration takes place. If they stay the same, reconfiguration
is not performed (optimization).

Signed-off-by: Darius Berghe <darius.berghe@analog.com>