drivers:adc:max11205: Add driver for max11205 device
The MAX11205 is compatible with the 2-wire interface and uses
SCLK and RDY/DOUT for serial communications. In this mode, all
controls are implemented by timing the high or low phase of the SCLK.
The 2-wire serial interface only allows for data to be read out through the
RDY/DOUT output.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>