iio: Add cyclic buffer support
Add cyclic buffer support when cyclic flag is given in
OPEN command. The cyclic flag is kept until a CLOSE command
is received.
When a device is opened in cyclic mode, the first buffer of samples
pushed by the client, will be pushed continuously to the iio
driver. submit function will be called continuously with the
given samples.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>