util.tcl: specify lwIP Echo Server template
In order to add the networking functionality to no-OS, lwip and
low level PHY files must be built into the application.

Vitis comes with an lwIP Echo Server template application that
we can use to accomplish this.

This commit changes the build scripts to be able to start an
application based on the lwIP Echo Server template.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>