maxim: max32655: Reconfigure SPI before transfers
Configure the SPI peripheral for the MAX32655 before a transfer
if the CS number is different from the last one. This done because
there could be multiple devices on the same SPI bus, which may
have different parameters (such as mode and maximum speed).

Replace the max_spi_write_and_read() code with a max_spi_transfer()
call (since that's just a particular case of transfer) to avoid
duplication.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>