drivers: temperature: adt75: ADT75 driver
The ADT75 is a complete temperature monitoring system in 8-lead MSOP
and SOIC packages. It contains a band gap temperature sensor and a
12-bit analog-to-digital converter (ADC) to monitor and digitize the
temperature to a resolution of 0.0625°C. The ADT75 is pin and register
compatible with the LM75 and AD7416.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>
(cherry picked from commit 6456ede335305e84a8d7adf505dd070053a1102b)